Update Solocraft.sql
from tinyint to bigint ( every Character with GUID <255 can scale endless)

+++ data/sql/db-characters/Solocraft.sql
@@ -1,6 +1,6 @@
 -- Structure for table custom_solocraft_character_stats
 CREATE TABLE IF NOT EXISTS `custom_solocraft_character_stats` (
-  `GUID` tinyint unsigned NOT NULL,
+  `GUID` bigint unsigned NOT NULL,
   `Difficulty` float NOT NULL,
   `GroupSize` int NOT NULL,
   `SpellPower` int unsigned NOT NULL DEFAULT '0',
